feat: moved SDG/region selector component from data page to home page
