+++ .changeset/flat-singers-crash.md
@@ -1,0 +1,5 @@
+---
+"geohub": patch
+---
+
+feat: moved SDG/region selector component from data page to home page
